+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Added [DJG103](doc/checks/DJG103.md) to look for use of the Django `extra()` API on a query set with quoted parameters.
 * Extended [DJG102](doc/checks/DJG101.md) to include quoted templated variable names, e.g. `'%(variable)s'`
 * Added [DJG104](doc/checks/DJG104.md) to inspect usage of the Django Expression APIs with dangerous templates vulnerable to SQL injection.
+* Modified [TRY100](doc/checks/TRY100.md) to only verify `Exception` and no exception specified.
 
 ## 1.11.0